requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
aiofiles==23.2.0
google-cloud-vision==3.4.5
python-dotenv==1.0.0
pydantic==2.5.0
gunicorn==21.2.0 